🧪 [testing improvement] Add comprehensive tests for AWS SigV4 implementation
